Print Method for rmd_doc

Quick display for rmd_doc objects. This method also defines the way how objects are displayed in the console.

Usage

# S3 method for rmd_doc
print(x, maxlines = 15, ...)

Arguments

x

An object of class rmd_doc.

maxlines

An integer value indicating the number of lines used for the display. Longer documents will be truncated.

...

Further arguments passed among methods (not yet in use).

Value

A display of the resulting R-Markdown document in the console.

Examples

## Create a new document
my_document <- write_rmd(
  title = "A journey in rOpenSci",
  author = "Miguel Alvarez",
  output = "html_document",
  body = read_rmd(file.path(
    path.package("yamlme"),
    "taxlistjourney.Rmd"
  ))
)
#> The yaml-header will not be imported by this function. Use 'update()' to re-write the header.
my_document
#> ---
#> title: A journey in rOpenSci
#> author: Miguel Alvarez
#> output: html_document
#> 
#> ---
#> 
#> list()# document imported by 'read_rmd'
#> c("\n", "After a long review process, where a lot of improvements have been implemented,\n", "the package [`taxlist`](https://docs.ropensci.org/taxlist/) managed to be\n", "accepted in [**rOpenSci**](https://ropensci.org/).\n", "The package `taxllist` handles taxonomic lists as **S4** objects in\n", "[**R**](https://www.r-project.org/).\n", "Such objects can be used as modules within data sets of biodiversity records as\n", "in the package [`vegtable`](https://github.com/kamapu/vegtable).\n", "\n", 
#> "In a brief talk a summary of the experiences during the process of submission to\n", "**rOpenSci** was presented in the context of the conference [**LatinR\n", "2020**](https://latin-r.com/en).\n", "Here you can access to the\n", "[slides](https://kamapu.github.io/documents/Alvarez2020.pdf) and the\n", "[video](https://www.youtube.com/watch?v=NDB4s8N2fTw) of the presentation.\n", "\n", "[![](http://img.youtube.com/vi/NDB4s8N2fTw/0.jpg)](http://www.youtube.com/watch?v=NDB4s8N2fTw \"\")\n")
if (FALSE) {
## Render the document
render_rmd(my_document, "taxlistjourney")
}