Print Method for rmd_doc

Quick display for rmd_doc objects. This method also defines the way how objects are displayed in the console.

# S3 method for rmd_doc
print(x, maxlines = 15, ...)

Arguments

x
An object of class rmd_doc.

maxlines
An integer value indicating the number of lines used for the display. Longer documents will be truncated.

...
Further arguments passed among methods (not yet in use).

Value

A display of the resulting R-Markdown document in the console.

Examples

## Create a new document
my_document <- write_rmd(
    title = "A journey in rOpenSci",
    author = "Miguel Alvarez",
    output = "html_document",
    body = read_rmd(file.path(path.package("yamlme"),
            "taxlistjourney.Rmd")))
my_document

#> ---
#> title: A journey in rOpenSci
#> author: Miguel Alvarez
#> output: html_document
#> ---
#> 
#> 
#> After a long review process, where a lot of improvements have been implemented,
#> the package [`taxlist`](https://docs.ropensci.org/taxlist/) managed to be
#> accepted in [**rOpenSci**](https://ropensci.org/).
#> The package `taxllist` handles taxonomic lists as **S4** objects in
#> [**R**](https://www.r-project.org/).
#> Such objects can be used as modules within data sets of biodiversity records as
#> in the package [`vegtable`](https://github.com/kamapu/vegtable).
#> 
#> In a brief talk a summary of the experiences during the process of submission to
#> **rOpenSci** was presented in the context of the conference [**LatinR
#> 2020**](https://latin-r.com/en).
#> Here you can access to the
#> [slides](https://kamapu.github.io/documents/Alvarez2020.pdf) and the
#> [video](https://www.youtube.com/watch?v=NDB4s8N2fTw) of the presentation.
#> 
#> [![](http://img.youtube.com/vi/NDB4s8N2fTw/0.jpg)](http://www.youtube.com/watch?v=NDB4s8N2fTw "")
#> 

if (FALSE) {
## Render the document
render_rmd(my_document, "taxlistjourney")
}